ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Agentic AI document verification service — 6 specialized agents with multi-tool self-correction loops."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "openai>=1.30.1",
    "pydantic-settings>=2.2.1",
    "python-json-logger>=2.0.7",
    "unidecode>=1.3.8",
    "pyarabic>=0.5.1",
    "opencv-python>=4.10.0",
    "numpy>=1.26.4",
    "google-genai>=1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
